+++ Client/src/App.css
@@ -1,6 +1,7 @@
 #root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
+  margin: 0;
+  padding: 1rem;
   text-align: center;
+  width: 100%;
+  height: 100vh;
 }
