+++ Client/src/App.jsx
@@ -2,9 +2,9 @@ import "./App.css";
 
 import { Route, Routes, useLocation } from "react-router-dom";
 import NavBar from "./components/NavBar/NavBar";
-import HomePage from "./Components/Home/HomePage";
 import Footer from "./Components/Footer/Footer";
 import CreateUserForm from "./components/createUserForm/CreateUserForm";
+import HomePage from "./components/HomePage/HomePage";
 
 function App() {
   const {pathname} = useLocation();
