--- Client/src/components/Home/HomePage.jsx
@@ -1,13 +1,0 @@
-import Hero from "../Hero/Hero"
-import Cards from "../Cards/Cards"
-
-const HomePage = () => {
-  return (
-    <div>
-        <Hero />
-        <Cards/>
-    </div>
-  )
-}
-
-export default HomePage
