+++ Client/src/components/HomePage/HomePage.jsx
@@ -1,0 +1,13 @@
+import Hero from "../Hero/Hero";
+import ProductList from "../ProductList/ProductList";
+
+const HomePage = () => {
+  return (
+    <div>
+      <Hero />
+      <ProductList />
+    </div>
+  );
+};
+
+export default HomePage;
