+++ Client/src/components/ProductList/ProductList.jsx
@@ -1,8 +1,9 @@
 /* eslint-disable react/prop-types */
 // ProductList.jsx
-import React, { useEffect, useState } from 'react';
+import  { useEffect, useState } from 'react';
 import { getAllProducts } from '../../redux/actions/actions';
-import s from './cards.module.css';
+import s from './ProductList.module.css';
+import Card from '../card/Card';
 
 const ProductList = () => {
   const [products, setProducts] = useState([]);
@@ -25,25 +26,12 @@ const ProductList = () => {
       <h1>Lista de Productos</h1>
       <div className={s.productList}>
         {products.map((product) => (
-          <ProductCard key={product.id} product={product} />
+          <Card key={product.id} product={product} />
         ))}
       </div>
     </div>
   );
 };
 
-const ProductCard = ({ product }) => {
-  const { name, price, colour, image } = product;
-  console.log(image)
-
-  return (
-    <div className={s.productCard}>
-      {image && <img src={`https://${image}`} alt={name} className={s.productImage} />}
-      <h3>{name}</h3>
-      <p>Precio: {price}</p>
-      <p>Color: {colour}</p>
-    </div>
-  );
-};
 
 export default ProductList;
