+++ Client/src/components/card/Card.module.css
@@ -1,0 +1,19 @@
+.productCard {
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  padding: 16px;
+  margin: 16px;
+  width: 200px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  transition: transform 0.3s ease-in-out;
+}
+
+.productCard:hover {
+  transform: scale(1.05);
+}
+
+.productImage {
+  height: 200px;
+  object-fit: cover;
+  border-radius: 10px;
+}
