+++ Client/src/components/searchbar/SearchBar.module.css
@@ -1,0 +1,12 @@
+.container{
+  display: flex;
+  justify-content: flex-end;
+  width: 30%;
+  height: 40px;
+}
+
+.input {
+  padding: 10px;
+  border-radius: 10px;
+  margin-right: 10px;
+}
